Make cache less buggy
The recent queue was changed to a guava Cache, and if the class
can't be found in the cache, it gets transformed again.
This fixes crashing on startup that happened.

Also, classes are placed in the recent cache even after being
transformed, this could cause speedups even when the cache
hasn't been built yet.